Refactor the caching code and add one more materialized view to cache.
Turns out the format of the results and general processing is very similar between the two types of materialized views. This refactoring merges the common factors and each will differ in the type and the sql query being used. The new type added is UnignoredTraces.

Bug: b/374764265
Change-Id: If6833e721cb126f43849cd0d03706c3b4235d1a1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/914847
Reviewed-by: Hao Wu <haowoo@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>