[perf] add legend as a side-panel in google chart
This CL turns off the default google chart legend and places the
content in a side panel. The side panel can be toggled on and off.
When there is only one trace, the side panel will not be shown.

The side panel can be adapted in the future to show other content
such as the toolbar. A scroll bar is used for now for legend content
that exceeds the viewport, but can also consider a drag feature to
make it wider or thinner.

Bug: b/370804689
Change-Id: Iac0f4b60ec39c89db4673b0844005046e642a004
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/916135
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>