Use alphanumeric keys to match parameters to pick links.
To create a trace ID, the keys are normalized for use in a URI. The trace ID is then parsed to retrieve these keys. Comparing the normalized keys with the original parameters caused a mismatch. Removing the symbols makes comparing the keys more reliable.

Fixed: 375686651
Change-Id: I84d278639c3021c5c2415b6bcbba456dbb206c14
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/916848
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>