Update sql schema generation to generate schema for Spanner
go/sql/exporter.go has the main change to support exporting the schema structs to a spanner compatible schema. The main changes are
- STRING -> TEXT
- UUID -> TEXT
- BYTES -> BYTEA
- gen_random_uuid() -> spanner.generate_uuid()
- Replace unique_rowid() with sequences per table.
- Index creation is a separate query.

This change only impacts the schema generation. SQL query changes will be done in a forthcoming CL.

Also updated the */tosql code for all the services using CDB to generate the schema for spanner as well. This is not being used anywhere yet.

Bug: 377531543
Change-Id: I80737765f2322a0f3ccbae1b2a21ee6b16cc1e0a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/916263
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>