[Triage Workflow] Add Checkbox logic for Anomaly Table.
When clicking on the header checkbox, the rest of the boxes will be
modified accordingly. When click the top checkbox of a group when
the group is collapsed, all the checkboxes in that group will be ticked.

Added Triage button that will show a pup up with the triage menu. This makes it so that we only update the triage menu with anomalies when the Triage button is clicked.

Bug: b/374810808

Change-Id: I5ff4a4daded0f8dfd1d9142880f7473c4911d221
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/917723
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>