Remove the graph from the list of graphs after closing.
Currently, closing a graph leaves a gap in the view and doesn't remove the graph object from the multigraphs list. This change saves a reference to the graph in the state. When the user closes the graph, it notifies `explore-multi-sk.ts` to remove it from the list.

Fixed: 377639959
Change-Id: Ic64d2cd5e1288c1ed40949a5b6a2ee506219a9fd
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/917789
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>