[perf] No wrap text on test-picker-sk
Set white-space: nowrap on buttons for test-picker-sk, which makes
the user need to scroll right to view the remainder of the page
when it gets too long, or when the window is too small.

This is in alignment with the way the chart works below.

Bug: chromium:380129895
Change-Id: I9d3794ca3bae723746cb83938a966b9acea4f462
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/926864
Reviewed-by: Vidit Chitkara <viditchitkara@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>