Hide the entire params/details table
Sufficient information is available through both the title and
the tooltip.

Hide the entire details tab when both title and tooltip are enabled.
If only tooltip is enabled, hide the commit details.
Only having title enabled will keep it the same.

Because the links are still critical, it's pulled up above the tabs.

Bug: chromium:380218859
Change-Id: I7e27315014126e93037fb6b11782c4cb8ba9a9cb
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/926861
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Vidit Chitkara <viditchitkara@google.com>