[perf] bring back tables, and only exclude commit message
With the query button and title, it was assumed that the table
at the bottom was duplicate info and not needed. This does not
seem to be the case per feedback.

Only hiding commit message when chart tooltip is set.

Bug: b:380218859
Change-Id: I04ca89f9b9380e8741db2a9d883d11a8a7c3034f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/927943
Auto-Submit: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>