[perf] Bug fix: Users need to make 2 clicks for plotting on /m/ page
On the first click there is a no-op on the UI however API requests are
fired. The second click actually plots the graph. This is not an ideal
user experience.

The removeAll function is used at places apart from the "Remove All" and
the "X" buttons. Instead of removeAll function, the event to clear the
plot should only be fired when the user actually wants to close the
graph.

Bug: b/380215495
Change-Id: Ic00e488f3f6c4a2f2b190e67411b7c252794ec01
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/930436
Commit-Queue: Vidit Chitkara <viditchitkara@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>