[perf] do not use sid if only one anomaly is selected on Graph button
The current backend handles differently when one or multiple anomaly
keys are received. Sid will be generated only when multiple anomlies
are received.

On the implementation of the Graph button, we always rely on sid. So
when only one anomaly is selected, no sid is generated and the graph
button opens a page but loads nothing.

This CL is a quick workaround. We should update the backend handling
to keep the front end simple.


Bug: b/378185555
Change-Id: I8e21838782222faa9bcf98f34578b91c04494ce3
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/931242
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Maggie Dong <jiaxindong@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>