[perf] do not send event on mouse down unless shift is held
This is Google chart specific.

On mouse down, we used to emit an event. This hid the tooltip, but
not in a correct manner - you can put the mouse over the selected
point and the tooltip would reappear.

This change makes it so that the user is required to click close.
By doing so selections on both the chart and the plot summary
are reset correctly.

When the tooltip is selected, meaning it's focused the user can
hold shift to figure out the delta.

Change-Id: I9fcb40448801f1d00939c9f6e6d8aa3d523da0f9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/931780
Auto-Submit: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>