[perf] expect selected_keys from anomalies response from Chromeperf
Adding a new field for the response to support the use case when
'graph' button is clicked.

With the selected_keys, front end will know which anomalies are
originally selected and thus needed to be chekced in the report page.

It has dependency for the catapult change:
https://chromium-review.googlesource.com/c/catapult/+/6111474

Bug: b/383464022
Change-Id: I55e06c7f6bf6bb239d9d320ae9ab8f81aa2bcf12
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/931735
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Maggie Dong <jiaxindong@google.com>