[perf] clean up test name refering to config.
The invalid chars are defined in config as invalid_param_char_regex. If it is
not defined, the one in query.InvalidChar will be used.

The current clean up in anomaliesApi.go does not look into the config. This CL
updates the logic to refer to the config.

Bug: b/383913153
Change-Id: Ia6d0053a5553ef58a9902ccf2cb947d92babfc84
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/936581
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>