[perf] send POST request for commit data when selecting on google chart
Google chart implementation requires users to click on data points
on the older plot-simple-sk for the commit to update into
pointToCommitDetailMap. Only this way could it receive the additional
commit information.

This change adds the POST request as part of selection so that the
commit info gets passed to tooltip. However, setting in
pointTocommitDetailMap doesn't WAI here.

Also, adding a try/catch on selection.row. This is undefined when users
click on the same data point multiple times, so instead of
erroring we'll just console.log and ignore.

Change-Id: I5b150f4aaba9e6e7efba1ba4d141e3e8e5a03ccf
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/931897
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>