[perf] enable displaying point links on click
point links were being imported to chart tooltip, but it was never
being set. this change introduces function loadPointLinks() in
chart tooltip to load point-links-sk.

The call to loadPointLinks is not embedded within enableTooltip
because it's only set onClick, and the list of params for
enableTooltip would get too long.

Change-Id: I78e37cfe4302da10cc5621f45dafed89545c0f9d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/942337
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>