Ensure consistent use of view and table indices across function calls.
Now that we add the ability to hide traces, the view and table indices can differ, so we can't use view indices for table operations and vice-versa.

For colors, we assign set colors to each trace so that hiding traces doesn't affect a color's value due to an index change.

For displaying the tooltip, we translate the view indices given by events to table indices that can be used to retrieve internal data such as commit position and trace name.

Change-Id: I0ddad998077a6e045d4435b686122b2bd6fb6eac
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/944145
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>