When recording new updates, retain the existing `notification_id`.
When a new regression is detected and a new issue tracker is generated, the `notification_id` is recorded correctly in the database. However, if a new regression is detected for the same key, the cluster information is overwritten, and the `notification_id` is lost. The system should read the existing `notification_id` and update the cluster information while retaining the ID.

Bug: 393375231
Change-Id: I2de9da95bed1893bc5fcfb8f41ee4d6e477ad820
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/945941
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>