[perf] enable tracing for Pinpoint workflows
create an opentelemetry interceptor configured by the opentelemetry
trace provider with the GCP exporter configured such that it
can export the traces.

the trace bridge library ensures that all opencensus/trace calls
are still exported to GCP. Migration from the tracing.Initialize()
call to tracing.InitializeOtel() is not part of this change and
will be handled separately.

Bug: b/394870809
Change-Id: Id9a928939ecde14de699ac4742f123ad157fd708
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/949977
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>