Add more search data to cache for Gold
Currently we only cache the untriaged images. This CL adds the positive, negative and ignored traces as well.
In the search call, based on the options selected, we retrieve the data from the cache and combine the results
in memory. These four are mutually exclusive sets so we should not see duplicates in the results.

Bug: b/388595215
Change-Id: I0a2e8d9d306f5109ae64e8d69a7a89a40969f29b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/950497
Reviewed-by: Farid (Mojtaba) Faridzad <faridzad@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>