[Perf] Add ReverseKeyMap to keep track of param value update
When a perf data point is uploaded from Chromeperf to Skia, the test path data will be updated. Specifically, the 'invalid' chars will be replaced with '_'.

Currently, we still rely on the anomaly data on Chromeperf. Thus, we need to use the Skia test path to query on Chromeperf side for anomalies. However, there is no way to reverse the update because we don't know which invalid data it was before being replaced to '_'.

We will create a new table here to keep track of each pair of values before and after the replacement.

The one-pager:
https://docs.google.com/document/d/1zcNDwEkJi3SkjLwV1RrMjrGlHXE2fRyF4TFaC3RnAPM

Bug: b/383913153
Change-Id: Iffa5274b0d1ee723f6b134fb8510ae8646a022aa
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/948832
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>