[pinpoint] implement service support for pairwise execution
Pairwise execution in isolation (try) for Pinpoint service was not
implemented. This is the initial change to start that.

There are still quite a few features missing, ie/ patches, writing
to gerrit, and the connection from chromeperf to invoke this API
will be done separately.

Bug: b/391717876
Change-Id: If0de76adaf24a692b97bc3d45ee29f71b96f5302
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/952137
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>