Modify extendRange to ensure all tracesets always have same length.
When we do a frameRequest for a given range, if there are no values to
return for a trace, the data frame will not include that trace.
Currently, if the trace is not present in the response, we will not
concatenate anything to the current trace. This can cause traces to have
mismatched length leading to shifts in the Graph.

In this CL, if traces are missing from the response but the header is
getting expanded, we need to ensure we pad the current traceset with
missing values so that the datapoints don't shift back.

Bug: b/398217790
Change-Id: Ie74528b36d57988fa44acbde37efa28025f6632e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/955102
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>