[perf] update call for bisect to skia pinpoint bisect
This was initially implemented to go through pinpoint service defined
in chromeperf. There's no need to do that anymore, because we have
the skia based pinpoint implementation. This change routes it to
the bisection workflow through temporal.

One caveat is that the Pinpoint URL will not be accessible until
the temporal job is done, because of the writeback that's done at the
end. The user needs to take the Job ID and go to the Temporal page
to see the actual status of the job.

In Pinpoint, the rate limiter is dropped because catapult
is calling culprit finder now.

Bug: b/398911040
Change-Id: Ia6e658e0128068351b233f531855128e8df657b8
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/963256
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>