[perf] add API for creating Pinpoint A/B try job from perf
This api uses the legacy api/new endpoint rather than routing
through the chromeperf redirect. The primary use of this api
is to trigger tracing jobs. We should avoid supporting try jobs
on legacy pinpoint in general since we plan to deprecate it in
the future.

Bug: b/391784563
Change-Id: I831f30fe1e153e8b1e0de1e8f7a317b248a53432
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/963917
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>