[perf] create pinpoint-try-job-dialog for creating try jobs in Pinpoint
This dialog takes in the commit positions and trace strings to record
additional traces in Pinpoint benchmark runs. This dialog assumes that
other inputs stay the same from the chart-tooltip. The goal of this
dialog is to fulfill the debug trace feature request. Additional fields
typically used for try jobs are omitted.

Bug: b/391784563
Change-Id: Ifcf887949c505a026b1d54b02c8b36ff5a5224cf
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/964016
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>