[perf] connect request debug trace to tooltip
This CL exposes the Request Debug Trace button onto the tooltip.
The button opens a dialog that allows a user to start a Pinpoint try
job with additional arguments that exposes more tracing data.

The user can only change the commit position and the trace arguments.
This selection mimics the features available in the legacy perf.

Bug: b/391784563
Change-Id: I854875415c2ef056f6f2de22b1a10e197378ae86
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/966739
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>