[perf] fix API alert/notify/try exampleSend()
When we call exampleSend(), the cluster summary does not have
"keys", so when the user clicks "Test" on an alert for Individual,
the logic would fall into the StepFitGrouping if statement and
hit an Index OOB error because cl.Keys is nil.

This adds a nil check to ensure that there are Keys even being set
before we try and go get the links.

Change-Id: Iecae9ba44af1ae5934da5cd3b2c4811d0738a746
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/968137
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Auto-Submit: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>