Revert "[Trace Name Schema] Disable preflight with default_param_selection"
This reverts commit 3b8b0135fd52117dd6b6b34a866d747004a205db.

Reason for revert: The cache keys are likely generated with the
assumption of stat=value, so when we try to match the first two
tiers, there's no match because we pulled stat out. This change
might be the root cause for slowness in production; I'm assuming
that the cache hits are not happening and we're querying.

Original change's description:
> [Trace Name Schema] Disable preflight with default_param_selection
>
> Context:
> go/trace-name-schema-changes lists that cache for handling
> level1_cache_key and level2_cache_key should ignore
> default_param_selections.
> Bug: b/404337891
> Change-Id: I5ce4ed97166586c16688c7f2989d0e6e4f2c6ed1
> Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/966737
> Commit-Queue: Arthur Wang <wuwang@google.com>
> Reviewed-by: Ashwin Verleker <ashwinpv@google.com>

Bug: b/404337891
Change-Id: Idf9aa0a789ab6ab6907798c02349d76a5785e3b2
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/968478
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>