[perf] clean up y-axis title and consolidate closeTooltip logic
This CL makes the google chart y-axis title into a property, which
will trigger an automatic update whenever the title is updated.

Thie CL consolidates the closeTooltip function into one place and
ensures consistent behavior across all user journeys except one edge
case where a user clicks on one data point and then clicks on a
separate data point. In theory we would like for the tooltip to
instantly enable again but because of how the update cycles work,
the select action is swallowed. This journey may require further
revision if users do not want to treat the graph as unaccessible when
the tooltip is up.

Bug: b/400335223
Change-Id: Iade77fc70c14ccf5ca04d08903eb99d9bd9b437f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/969758
Commit-Queue: Tony Seaward <seawardt@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>