[perf] add text to tooltip to support android use case
This CL adds text keys to the point-links-sk. The text keys will
show the value upon hover and are accompanied with a button to copy
if a user needs to access the info directly.

The CL also revises the androidx configs to look for these new
text keys.

Bug: b/372515172
Change-Id: I2c631435e098a54a278ff90e2115eef639b7302a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/971161
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>