[pinpoint] refactor culprit logic out of pairwise
This CL moves the logic that determines if a regression is significant
out of the pairwise job and into the culprit_finder workflow. This CL
ensures skia Pinpoint has the same features as legacy Pinpoint while
allowing try jobs to be run without the chart as input.

Bug: b/406763900
Change-Id: I02545612eb9da6cb535d26f80a65fa05737521f9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/988398
Reviewed-by: Eduardo Yap <eduardoyap@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>