Fixed the inconsistent test path issue
The test path in the bisect dialog was inconsistent with the selected trace name. This was because when fetching the cid, the rest of the code executed without waiting for the api call's response. Updated the call to be asynchronous.
Removed duplicate loadpointlink method call.

Bug: b/416179817

Change-Id: I705f1acd715ffa37015be1fdb9e25b13139520b1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/994101
Auto-Submit: Maggie Dong <jiaxindong@google.com>
Commit-Queue: Maggie Dong <jiaxindong@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>