Migrate database to new version in periodic tasks
This is a reland of https://skia-review.googlesource.com/c/buildbot/+/992260 with some changes

- json files have been manually changed to reflect the current prod db schema
- no db migration queries have been specified


Once this CL gets deployed in production, the expectation is for periodictasks.go to not throw any error. Once that is done, we would know for sure that json files are consistent, and later CLs can contain right migration statements.

Change-Id: If1393a67e68dded74500508907f38b6457e09ea7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/997076
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Prakhar Asthana <pasthana@google.com>