Fixed an issue that caused the legend to not display when only a single valid trace was present.
The issue is that when there are two traces - one being a valid trace and the other a non-valid trace with no data (trace name is "special_zero"). The "special_zero" trace needs to be removed. Otherwise it causes the filtering of unique keys for legend generation to fail.


Bug: b/415094970

Change-Id: I23d730e15401f4059b5a3b9ff8070946bf5db85b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1000180
Commit-Queue: Maggie Dong <jiaxindong@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>