[logging] Use Cloud Logging package for structured logging
The RedirectAsJSON option causes all logs to be emitted to the given
file as JSON-structured logs instead of sent via the Cloud Logging API.
This is equivalent to what this package was already doing but uses the
Cloud Logging package's built-in structs to avoid accidental mismatches.

Unfortunately, log splitting isn't implemented or supported via the
client library or API itself.  Implement splitting of log entries based
on the presumed length limit for log messages.  We won't be able to
group these entries in the UI, but at least we'll have fewer of them to
scroll through.

Change-Id: I5ffb6d38c2a7967ff336866366e5f79ab6ae697f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1003256
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>
Commit-Queue: Eric Boren <borenet@google.com>