[perf] add mcp chromeperf tools for benchmark, bots, stories
add tooling for listing benchmark, bots and stories.
for listing bots, use /api/describe if the benchmark is provided so that
the results are tailored. otherwise, use /api/config.
this introduces a lightweight chromeperf client struct that does not
use any existing client implementations in this repository to be able to
call the other APIs and be able to test them in isolation. this can
be revisited to create a single client implementation should we
want to expand the API set for other use cases.
this change also introduces a common/ folder containing some logic that
is shared between chromeperf and pinpoint, such as content type header
and a default http oauth client.

Change-Id: Ia0aee882391d5cea2ce0f8221c3051b00818e4ff
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1004816
Reviewed-by: Nelson Billing <nbilling@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>