Implement chromiumbuilder Shutdown
Implements the Shutdown function for the chromiumbuilder MCP service.
This primarily looks to avoid killing git-related calls, which can
potentially leave the checkout in a bad state that could affect future
deployments.

In the one case where we cannot avoid killing git mid-run (while
fetching Chromium), the affected directories are wiped to give the next
deployment a clean slate.

As a result of this, a number of synchronization members are added to
the service to keep track of what is currently happening when the
server is shut down.

Change-Id: I7f1d30f9016799ee132c13b38726de48e75dfcb7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1006318
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>