Temporary fix: The user was not able to add an issue.
The reason is that the previous traces are stored in format with commas(e.g ,a=1,) in the user issue sql database. However, we recently removed the first and last comma for special function cases. So it's unable to get or save new user issues because it can't find the trace path.

Next step is to update user issue map data in the sql database.

Bug: b/424208820

Change-Id: I49bdcceed66eee783582fef975f1bdd86f24488d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1007197
Commit-Queue: Maggie Dong <jiaxindong@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>