Enforce no concurrent chromiumbuilder requests
Updates the chromiumbuilder MCP service to only allow one tool request
to be handled at a time via a mutex that is acquired any time a request
is handled. This was not necessary when running locally, but now that
the service is deployed, it may be possible for multiple clients to
issue concurrent requests. Since the service is backed by a single
depot_tools/Chromium checkout, multiple requests would trample on each
others' work.

This is not a good long-term solution, but should prevent any issues for
now.

Change-Id: Iec7ded0b07351b2a862e94cec16150934e858b6d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1007917
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>