Add chromiumbuilder CI child tester tool
Adds a new tool to the chromiumbuilder service to create a new child
tester for an existing parent builder. As a result, a fair amount of
code is refactored to be reusable between the two tools, but there
is no functional change as a result.

Change-Id: I5bdae421e8e258ef9202de9341071959108103dc
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1008224
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>