[Pinpoint] Implementing JobStore object
These specifically defined functions will be used primarily within the
Pairwise.go workflow orchestrater to write back to the Spanner database
when encountering details of a job.

More functions could be implemented in the future to write back more
information to the database.

job_name & submitted_by are fields in the database that would be
populated by the frontend Pinpoint service as a means to populate the
landing page with identifying details. These values aren't propagated to
the backend and really only are dealt with the frontend and auth.

Bug: crbug.com/336330069

Change-Id: Iac69b2d48ccb22f26a84bfea5f04c7bc07370733
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1015178
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>