[Pinpoint] Modules for listing jobs and scaffold for the Pinpoint Frontend. API typescript handler for the ListJobs endpoint.
This is a first in the series of CLs for the new Pinpoint UI. This is
for the landing page and these modules are meant to be the children of a
larger module that deals with the server side logic for the landing page
specifically.

I wanted to try my best to split the project into multiple CLs for
submission.

Change-Id: If462df5b1032bdb46ebe3932879b13ef7c9fe0ba
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1023125
Reviewed-by: Maggie Dong <jiaxindong@google.com>
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>