[Pinpoint] Update Service handler logic to serve landing page HTML.
This is part of a larger code commit for implementing access to the
landing page of the new Pinpoint UI. This is my attempt at breaking down
the code contribution to smaller manageable chunks.

Context: https://skia-review.googlesource.com/c/buildbot/+/1023125 deals
with smaller children modules that will build off a
pinpoint-landing-page-sk module that will serve as a larger orchestrator
of the landing page. That module will be added next while this CL deals
with setting up the entry point for the soon to be module.

landing-page.html will import pinpoint-landing-page-sk and break down
into the children modules.

Change-Id: I18e100bd90b5bc0990c3e6709fb99fc277e862f7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1023201
Reviewed-by: Maggie Dong <jiaxindong@google.com>
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>