Update pinned version of alpine linux to v3.22
https://louhi.corp.goog/6316342352543744/artifact/5597158639534080

We'll want to let this bake in for a bit, as a few services
like auth-proxy and temporal use this.

After this does, I think we can update jsfiddle, skottie,
and shaders to use this slimmer image.

Note the previous version was 3.9 which is old
https://alpinelinux.org/posts/Alpine-3.9.0-released.html

Bug: b/432555058
Change-Id: I353d7db729ce70da5c514a90cc797f4af2ebd79f
No-Try: true
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1024997
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>