[perf][UI] More straightforward update logic for PlotSummarySk
I was able to reproduce the issue with graphs not loading, and the only
thing that looked suspicious was a warning: "plot-summary-sk.ts:104 The
dataframe is not assigned because the element is not ready. Try call
await this.updateComplete first.". This looks like a data race.

To simplify the execution order here, I've made updateDataView
synchronous and triggered by updated (called after the update) instead
of willUpdate (called before the update). This ensures state consistency
during each update and should be logically equivalent.

Bug: 430254670
Change-Id: I7f64f370d727d835337fda6b15ec041a46fbcc4e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1024957
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Anri Sidorov <ansid@google.com>