[Pinpoint] Implement basic New Try Job Dialog for UI
This current implementaiton uses the benchmark, bot, and story API to
populated a Dropbox for a user to choose from. Stories can only be
chosen after a benchmark is chosen. And when a benchmark is chosen it
will list new bots that can run the given benchmark.

There is a simplified table that will be implemented in the future with
more defaulted option for casual users.

Here are visuals taken from locally hosting the modal: Base Page:
https://photos.app.goo.gl/6qHLSYZUxm413CuP9 Second Half of the page:
https://photos.app.goo.gl/7czBEE8YQL2aqPtZ8 Dropbox:
https://photos.app.goo.gl/G6dtuDqSbAv1CxbR7

Change-Id: Ib8d2e72cb34785bc49e8eec48507fd4c45569599
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1026896
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>
Reviewed-by: Tony Seaward <seawardt@google.com>