[gold][client] Add combine-inexact-matches option
Adds the combine-inexact-matches optional test argument to goldctl. If
this is set to 1 when performing an inexact image comparison, a
successful image comparison will cause the known good image that was
compared against to be reported instead of the newly produced image.

This is intended to be used for tests which are known to produce many
similar images. Reporting and triaging all of these images can result
in a bloated positive digest list. In the concrete example this is
being added to solve, the current behavior resulted in ~15k images
being triaged in the span of a month from a single test.

As a side effect of this, code has been refactored so that the
pass/fail logic is handled by helpers. This keeps the new code path
distinct from the standard code path. In addition to being subjectively
more readable, this was also necessary in order to not regress
performance in the standard code path, as the standard path is able
to upload the result JSON asynchronously much earlier.

Bug: b/434910221
Change-Id: I550cabc240fb1bf4da0582465fa9240c060effee
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1031814
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>