[Pinpoint] Allow for a Pinpoint Try job to be initiated in the frontend.
To accomplish this a few changes needed to be made.

Backend: The core workflow logic has been refactored to use a stable jobStoreId for all database operations. Previously, the workflow would generate a new ID, which could cause issues with database indexing and tracking job status. Now, the jobStoreId is created upstream and passed into the workflow, ensuring a consistent identifier is used for all related database interactions

Frontend: pinpoint-new-job-sk provides a form to schedule a Pairwise job. This UI component validates user input and makes a call to the backend's gRPC API to initiate the job. The API has been updated to use a JSON body for the SchedulePairwiseRequest.


Change-Id: I481678b3fa8b8199bee7222c641f525d4d4bf17b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1033319
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>