[task scheduler] Add periodic repo update as backup to pubsub
Right now we're relying on nack() to resend the pubsub message when
syncs fail. I think there are a couple of problems with this:

1. When we take a long time to sync, pubsub might try to deliver the
   message again, which compounds the problem by adding duplicate syncs.
2. When pubsub eventually backs off trying to deliver the message, we
   stop trying to perform the sync.

This change adds a periodic update of each repo in addition to the
pubsub-triggered update and ensures that we're only trying to perform
one update of a given repo at a given time (not counting try jobs).

Bug: b/376865758
Change-Id: I3fccd1f505420abe558a086495dca7098383d7af
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1035736
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>