[Pinpoint] Implement Wilcoxon results page module
For now we implement a defaulted improvement directions (whether higher
is better or if lower is better) variable for the benchmarks we
currently support. A larger TODO would be to incorporate improvement
direction values for each of the benchmarks we support.

The source of truth for this fact isn't as clear and might require some
hard coding for each benchmark.

Here is a current mock: https://photos.app.goo.gl/HbcC4waaAncvTGkSA

Change-Id: I1daa07a1ec9fcc58e51816664fbf14a90ee5ea65
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1034751
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>