[Pinpoint] Add date filtering capabilities for UI
This change introduces the ability to filter Pinpoint jobs by a creation
date range on the jobs list page. Users can now specify a start and end
date to narrow down the displayed jobs.

The backend handler for the jobs list API has been updated to accept
start_date and end_date query parameters. It validates that both dates
are provided together, are in the YYYY-MM-DD format, and that the start
date does not come after the end date. The SQL query in the jobs store
has been modified to include createdat >= and createdat <= clauses based
on these new parameters.

Change-Id: I34f265111e939e1ea7de8cbbe84f0e8f86d515a8
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1037039
Reviewed-by: Tony Seaward <seawardt@google.com>
Commit-Queue: Natnael Alemayehu <natnaelal@google.com>