[perf] Fix flaky dfbuilder tests.
If the commit is missing then req.commitNumberToOutputIndex returns 0, so the
first value of the response trace gets overwritten.

Fix is to add explicit test for map membership (ok).

This bug made the test flaky because it was a race condition on which of the two
tiles was handled by the merge worker first. If the second tile was handled
second, then the first value in the resulting trace would get overwritten with
the wrong value.

Change-Id: I95b442c57847fb2f4864a5c16bb55bfa1338cee1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1046756
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>