[perf] Change fatal error to error log when querying traceids.
This prevents the program from crashing when there's an error querying trace IDs, instead logging the error and allowing the program to continue.

Originally this was set to Fatal, because the error is never supposed to happen in this code (it's fully local). But: we shouldn't crash even in the case of bugs, and also the context can be cancelled. Fatal logs are discouraged to be used on prod.

Bug: 446827886
Change-Id: Id4ce1704e2f71fa3ca1aac3c4b9d5756d524bc9c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/1061196
Reviewed-by: Marcin Mordecki <mordeckimarcin@google.com>
Commit-Queue: Anri Sidorov <ansid@google.com>
Commit-Queue: Marcin Mordecki <mordeckimarcin@google.com>